Add a 'Customers' model
The model has two fields.

An email address field as a primary key which uses Django's email verifier.

A CharField for the customer's name.
I might separate the field into 'first_name' and 'last_name',
but I need to figure out a way to combine them in the __str__ method
so that the full name is returned.

I've also implemented a __str__ method which returns the name field.