Rename the 'Customers' model to 'Customer'
I realised that the admin page for Django automatically pluralises
the name of models for the tables.

It also is a better name from a type perspective, as the class which
represents the customer does not model multiple customers in of itself,
it represents a single instance of a customer.