Add an 'Order' model
This model has a foreign key on the customer's table, allowing orders
to be associated with customers.

It also contains a date & time field and a table number.
The table number has to be unique for a given date & time value.
This helps to prevent double bookings for a table.