Change the name field of 'Customer' to first_name & last_name
Remove the name field and replace it with two string fields explicitly
for the customer's first and last names.

The fields are half the length of the original field
at 15 characters each.

The __str__ method has also been rewritten to return the first and last
names, separated by a space.