Add a table showing the orders in the database to the 'show_orders' page
In the case that the database is empty of orders, the page does not
display a table but a message alerting the reader of this.

When displaying the orders in a table, the details of the order are
shown across three columns, the customer who placed the order,
the date and time the order is booked for, the table number booked.