Clear and recreate the django migrations
I deleted all the previously created migrations and had
django create a new one for the models.

I did this because the previous migrations were facing an issue with
the 'User' model when creating a new 'Order' record.

I believe that this was due to previously created and deleted records
not having a unique value for the id of the 'User' property.

When migrating the changes to the model, I set the NULL fields which
did not have a value for this property to '1'.

Clearing the previously created migrations, deleting the old database,
and then having django create the necessary migrations and datebase
solved this issue.