Change the django user model to my 'Customer' model
I decided that I didn't want the user to have to go through two
separate stages for creating an account, one for a user account,
created using django's user model, and another for the customers
table to add their orders against.

So, I made django use my 'Customer' model as the user model.

In doing this, I've removed the 'user' field from the model and
added a 'username' field, which just accepts a string.

I've configured the 'first_name', 'last_name' and 'email' fields to
be required by a new record in the table.

I also removed the previous migrations and re-created the migrated
django code; just to avoid any nasty issues with migration.